fix double notification
notifications come from initial subscription and from change on the
object. the race was possible. now completion of publishing comes when
object is loaded and it is equal to the published.
assume that there is only one publisher for the same object, which was
always the rule of the game.